refactor: expose header for user defined callback use RcppModule to implement split_callback
