+++ src/Makevars
@@ -1,3 +1,3 @@
 CXX_STD = CXX11
 
-PKG_CPPFLAGS =	-I../inst/include/
+PKG_CPPFLAGS =	-I../inst/include/ -DNOISY_DEBUG
