+++ src/converters.h
@@ -16,16 +16,12 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef __VECTOR_CONVERTER_HPP__
-#define __VECTOR_CONVERTER_HPP__
+#ifndef __CONVERTERS_HPP__
+#define __CONVERTERS_HPP__
 
-#include "hash_function.h"
+#include "vector_converter.h"
 #include "split.h"
-#include "callback.h"
-#include <Rcpp.h>
 
-struct VectorConverterParam;
-class VectorConverter;
 class CharacterConverter;
 class FactorConverter;
 template<typename ValueType, int RType>
@@ -36,7 +32,6 @@ class TagCountFactorConverter;
 class TagCountCharacterConverter;
 class InteractionConverter;
 
-typedef VectorConverterParam Param;
 typedef std::shared_ptr<VectorConverter> pVectorConverter;
 typedef std::shared_ptr<CharacterConverter> pCharacterConverter;
 typedef std::shared_ptr<FactorConverter> pFactorConverter;
@@ -52,84 +47,6 @@ typedef std::shared_ptr<TagCountFactorConverter> pTagCountFactorConverter;
 typedef std::shared_ptr<TagCountCharacterConverter> pTagCountCharacterConverter;
 typedef std::vector< pVectorConverter > ConvertersVec;
 typedef std::shared_ptr<InteractionConverter> pInteractionConverter;
-
-/**
- * Paramter of initializing VectorConverter
- */
-struct VectorConverterParam {
-  
-  std::string name;
-  HashFunction* h_main;
-  HashFunction* h_binary;
-  size_t hash_size;
-
-  VectorConverterParam(const std::string& _name, HashFunction* _h_main, HashFunction* _h_binary, size_t _hash_size)
-  : name(_name), h_main(_h_main), h_binary(_h_binary), hash_size(_hash_size) 
-  { }
-  
-};
-
-class VectorConverter {
-
-protected:
-  std::vector<uint32_t> feature_buffer;
-  std::vector<double> value_buffer;
-  std::string name;
-  size_t name_len;
-  HashFunction *h_main, *h_binary;
-  size_t hash_size;
-
-public:
-
-  bool is_final;
-  
-  explicit VectorConverter(const Param& param) 
-  : name(param.name), name_len(param.name.size()), h_main(param.h_main), 
-  h_binary(param.h_binary), hash_size(param.hash_size), is_final(true)
-  { }
-  
-  virtual ~VectorConverter() { }
-  
-  /**
-   * Evaluate the hashed feature from the raw data.
-   * The \code{get_feature} should be called before \code{get_value}
-   */
-  virtual const std::vector<uint32_t>& get_feature(size_t i) = 0;
-  
-  /**
-   * Evaluate the value from the raw data
-   * The \code{get_feature} should be called before \code{get_value}
-   */ 
-  virtual const std::vector<double>& get_value(size_t i) = 0;
-  
-  const std::string& get_name() const {
-    return name;
-  }
-  
-protected:
-
-  /**
-   * @return -1 if (int) v < 0, or +1 otherwise
-   */ 
-  static inline int get_sign(uint32_t v) {
-    if ((int) v < 0) return -1;
-    else return 1;
-  }
-
-  uint32_t get_hashed_feature(HashFunction* h, const char* str) {
-    name.append(str);
-    #ifdef NOISY_DEBUG
-    Rprintf("hashing %s ... ", name.c_str());
-    #endif
-    uint32_t retval = (*h)(name.c_str(), name.size());
-    #ifdef NOISY_DEBUG
-    Rprintf(" got %zu \n", retval);
-    #endif
-    name.resize(name_len);
-    return retval;
-  }
-
-};
 
 class CharacterConverter : public VectorConverter {
   
@@ -540,56 +457,4 @@ private:
   
 };
 
-class CallbackConverter : public VectorConverter {
-  
-  const CallbackFunctor* f;
-  Rcpp::CharacterVector src;
-  SEXP psrc;
-  std::vector< std::string > cache;
-  
-public:
-  
-  CallbackConverter(const CallbackFunctor* _f, SEXP _src, const Param& param)
-    : f(_f), src(_src), psrc(_src), VectorConverter(param) 
-    { }
-  
-  virtual ~CallbackConverter() { }
-  
-  virtual const std::vector<uint32_t>& get_feature(size_t i) {
-    SEXP pstr = STRING_ELT(psrc, i);
-    if (pstr == NA_STRING) {
-      feature_buffer.clear();
-    } else {
-      const char* str = CHAR(pstr);
-      cache = f->operator()(str);
-      feature_buffer.resize(cache.size());
-      std::transform(cache.begin(), cache.end(), feature_buffer.begin(), 
-                     [this](const std::string& s) {
-                       return this->get_hashed_feature(this->h_main, s.c_str());
-                     });
-      if (is_final) std::transform(feature_buffer.begin(), feature_buffer.end(), 
-          feature_buffer.begin(), [this](uint32_t feature) {
-            return feature % this->hash_size;
-          });
-    }
-    return feature_buffer;
-  }
-  
-  virtual const std::vector<double>& get_value(size_t i) {
-    SEXP pstr = STRING_ELT(psrc, i);
-    if (pstr == NA_STRING) {
-      value_buffer.clear();
-    } else {
-      const char* str = CHAR(pstr);
-      value_buffer.resize(cache.size());
-      std::transform(cache.begin(), cache.end(), value_buffer.begin(), 
-                     [this](const std::string& s) {
-                       return this->get_hashed_feature(this->h_binary, s.c_str());
-                     });
-    }
-    return value_buffer;
-  }
-  
-};
-  
-#endif // __VECTOR_CONVERTER_HPP__
+#endif // __CONVERTERS_HPP__
