--- src/digest.c
@@ -1,23 +1,0 @@
-/*
- * This file is part of FeatureHashing
- * Copyright (C) 2014-2015 Wush Wu
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation, either version 3 of the License, or (at your option)
- * any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of  MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "pmurhashAPI.h"
-
-const uint32_t 
-  MURMURHASH3_H_SEED = 3120602769LL,
-  MURMURHASH3_XI_SEED = 79193439LL;
