+++ src/hashed_model_matrix.h
@@ -26,6 +26,19 @@
 #include "callback.h"
 #include "hash_function.h"
 #include "vector_converter.h"
+#include "converters.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+uint32_t PMurHash32(uint32_t, const void*, int);
+
+extern const uint32_t MURMURHASH3_H_SEED, MURMURHASH3_XI_SEED;
+
+#ifdef __cplusplus
+}
+#endif
 
 typedef std::map< std::string, std::string > NameClassMapping;
 typedef std::vector< std::string > StrVec;
